+++ client/package.json
@@ -8,6 +8,7 @@
     "@testing-library/react": "^13.4.0",
     "@testing-library/user-event": "^13.5.0",
     "axios": "^1.3.4",
+    "html-react-parser": "^3.0.16",
     "js-cookie": "^3.0.1",
     "react": "^18.2.0",
     "react-contenteditable": "^3.3.7",
