+++ client/src/pages/UserSettingsPage.js
@@ -8,6 +8,7 @@ import UserSettingsInterests from "../components/UserSettingsInterests";
 import UserSettingsProfilePicture from "../components/UserSettingsProfilePicture";
 import DeleteAccountModal from "../components/DeleteAccountModal";
 import ReactQuill from "react-quill";
+import quillModules from "../utils/quillModules";
 import "react-quill/dist/quill.snow.css";
 
 export const UserSettingsPage = () => {
@@ -23,21 +24,6 @@ export const UserSettingsPage = () => {
   const [readme, setReadme] = useState(user.readme || "");
   const [selectedItems, setSelectedItems] = useState(fields);
   const [isDeleteModalOpen, setIsDeleteModalOpen] = useState(false);
-
-  const quillModules = {
-    toolbar: [
-      [{ header: [false, 2, 3] }],
-      ["bold", "italic", "underline", "strike"],
-      [
-        { list: "ordered" },
-        { list: "bullet" },
-        { indent: "-1" },
-        { indent: "+1" },
-      ],
-      ["link"],
-      ["clean"],
-    ],
-  };
 
   const handleSubmit = async (e) => {
     e.preventDefault();
@@ -172,18 +158,6 @@ export const UserSettingsPage = () => {
           >
             README
           </label>
-          {/* <textarea
-            id="readMe"
-            rows="8"
-            cols="50"
-            placeholder={readme || "Tell us a little bit about yourself..."}
-            value={readme}
-            onChange={(e) => setReadme(e.target.value)}
-            className="border border-slate-900 rounded w-full py-2 px-3 text-gray-700 
-              leading-tight focus:outline-bluegray resize-none"
-            required={false}
-          /> */}
-
           <ReactQuill
             id="readMe"
             value={readme}
