+++ client/src/utils/quillModules.js
@@ -1,0 +1,16 @@
+const quillModules = {
+  toolbar: [
+    [{ header: [false, 2, 3] }],
+    ["bold", "italic", "underline", "strike"],
+    [
+      { list: "ordered" },
+      { list: "bullet" },
+      { indent: "-1" },
+      { indent: "+1" },
+    ],
+    ["link"],
+    ["clean"],
+  ],
+};
+
+export default quillModules;
